+++ crates/cli/tests/reference/generic-import.bg.js
@@ -41,7 +41,7 @@ export function __wbg_new_typed_cceaf62d8d95e9f2(arg0, arg1) {
             const a = state0.a;
             state0.a = 0;
             try {
-                return wasm_bindgen_0000000000000037___convert__closures_____invoke___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined_______true_(a, state0.b, arg0, arg1);
+                return wasm_bindgen_0000000000000043___convert__closures_____invoke___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined_______true_(a, state0.b, arg0, arg1);
             } finally {
                 state0.a = a;
             }
@@ -70,7 +70,7 @@ export function __wbg_run_ef366b557a6598c4(arg0, arg1, arg2) {
             const a = state0.a;
             state0.a = 0;
             try {
-                return wasm_bindgen_0000000000000037___convert__closures_____invoke___bool__true_(a, state0.b, );
+                return wasm_bindgen_0000000000000043___convert__closures_____invoke___bool__true_(a, state0.b, );
             } finally {
                 state0.a = a;
             }
@@ -116,18 +116,18 @@ export function __wbg_then_72819b8d4e081fb5(arg0, arg1, arg2) {
     return ret;
 }
 export function __wbindgen_generic_0000000000000000(arg0, arg1) {
-    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [Externref], shim_idx: 59, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
-    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000037___convert__closures_____invoke___wasm_bindgen_0000000000000037___JsValue__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_);
+    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [Externref], shim_idx: 71, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
+    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000043___convert__closures_____invoke___wasm_bindgen_0000000000000043___JsValue__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_);
     return ret;
 }
 export function __wbindgen_generic_0000000000000001(arg0, arg1) {
-    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [String], shim_idx: 60, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
-    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000037___convert__closures_____invoke___alloc_000000000000003a___string__String__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_);
+    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [String], shim_idx: 72, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
+    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000043___convert__closures_____invoke___alloc_0000000000000046___string__String__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_);
     return ret;
 }
 export function __wbindgen_generic_0000000000000002(arg0, arg1) {
-    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [U32], shim_idx: 61, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
-    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000037___convert__closures_____invoke___u32__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_);
+    // Cast intrinsic for `Closure(Closure { owned: true, function: Function { arguments: [U32], shim_idx: 73, ret: Result(Unit), inner_ret: Some(Result(Unit)) }, mutable: true }) -> Externref`.
+    const ret = makeMutClosure(arg0, arg1, wasm_bindgen_0000000000000043___convert__closures_____invoke___u32__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_);
     return ret;
 }
 export function __wbindgen_generic_0000000000000003(arg0, arg1) {
@@ -167,65 +167,93 @@ export function __wbindgen_generic_0000000000000008(arg0, arg1) {
     // generic import `attach`: [Ref(Externref), Ref(Externref)] -> Unit
     arg0.attach(arg1);
 }
-export function __wbindgen_generic_0000000000000009(arg0, arg1) {
+export function __wbindgen_generic_0000000000000009(arg0) {
+    // generic import `blockInheritedImplTrait`: [F64] -> Unit
+    blockInheritedImplTrait(arg0);
+}
+export function __wbindgen_generic_000000000000000a(arg0) {
+    // generic import `blockInheritedImplTrait`: [U32] -> Unit
+    blockInheritedImplTrait(arg0 >>> 0);
+}
+export function __wbindgen_generic_000000000000000b(arg0, arg1) {
     // generic import `blockInheritedTwo`: [U32, F64] -> F64
     const ret = blockInheritedTwo(arg0 >>> 0, arg1);
     return ret;
 }
-export function __wbindgen_generic_000000000000000a(arg0) {
+export function __wbindgen_generic_000000000000000c(arg0) {
     // generic import `blockInherited`: [F64] -> Unit
     blockInherited(arg0);
 }
-export function __wbindgen_generic_000000000000000b(arg0) {
+export function __wbindgen_generic_000000000000000d(arg0) {
     // generic import `blockInherited`: [U32] -> Unit
     blockInherited(arg0 >>> 0);
 }
-export function __wbindgen_generic_000000000000000c(arg0) {
+export function __wbindgen_generic_000000000000000e(arg0) {
     // generic import `a.b.deepLog`: [F64] -> Unit
     a.b.deepLog(arg0);
 }
-export function __wbindgen_generic_000000000000000d(arg0, arg1, arg2) {
+export function __wbindgen_generic_000000000000000f(arg0, arg1, arg2) {
     // generic import `delete_indexed`: [Ref(Externref), Ref(String)] -> Unit
     delete arg0[getStringFromWasm0(arg1, arg2)];
 }
-export function __wbindgen_generic_000000000000000e(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000010(arg0, arg1, arg2) {
     // generic import `fillSlice`: [RefMut(Slice(U16)), U32] -> Unit
     fillSlice(getArrayU16FromWasm0(arg0, arg1), arg2 >>> 0);
 }
-export function __wbindgen_generic_000000000000000f(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000011(arg0, arg1, arg2) {
     // generic import `get`: [Ref(Externref), Ref(String)] -> U32
     const ret = arg0[getStringFromWasm0(arg1, arg2)];
     return ret;
 }
-export function __wbindgen_generic_0000000000000010(arg0) {
+export function __wbindgen_generic_0000000000000012(arg0) {
     // generic import `identity`: [F64] -> F64
     const ret = identity(arg0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000011(arg0) {
+export function __wbindgen_generic_0000000000000013(arg0) {
     // generic import `identity`: [U32] -> U32
     const ret = identity(arg0 >>> 0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000012(arg0) {
+export function __wbindgen_generic_0000000000000014(arg0) {
     // generic import `kind`: [Ref(Externref)] -> U32
     const ret = GetOwnOrInheritedPropertyDescriptor(Widget.prototype, 'kind').get.call(arg0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000013(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000015(arg0, arg1, arg2) {
     var v0 = Array.from(getArrayU16FromWasm0(arg0, arg1));
     // generic import `logBlockSlice`: [Ref(Vector(U16)), U32] -> Unit
     logBlockSlice(v0, arg2 >>> 0);
 }
-export function __wbindgen_generic_0000000000000014(arg0, arg1) {
+export function __wbindgen_generic_0000000000000016(arg0, arg1) {
     // generic import `logGenericSlice`: [Ref(Slice(F64))] -> Unit
     logGenericSlice(getArrayF64FromWasm0(arg0, arg1));
 }
-export function __wbindgen_generic_0000000000000015(arg0, arg1) {
+export function __wbindgen_generic_0000000000000017(arg0, arg1) {
     // generic import `logGenericSlice`: [Ref(Slice(U32))] -> Unit
     logGenericSlice(getArrayU32FromWasm0(arg0, arg1));
 }
-export function __wbindgen_generic_0000000000000016(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000018(arg0) {
+    // generic import `logImplTrait`: [F64] -> Unit
+    logImplTrait(arg0);
+}
+export function __wbindgen_generic_0000000000000019(arg0) {
+    // generic import `logImplTrait`: [U32] -> Unit
+    logImplTrait(arg0 >>> 0);
+}
+export function __wbindgen_generic_000000000000001a(arg0, arg1) {
+    var v0 = getArrayF64FromWasm0(arg0, arg1).slice();
+    wasm.__wbindgen_free(arg0, arg1 * 8, 8);
+    // generic import `logNestedImplTrait`: [Vector(F64)] -> Unit
+    logNestedImplTrait(v0);
+}
+export function __wbindgen_generic_000000000000001b(arg0, arg1) {
+    var v0 = getArrayU32FromWasm0(arg0, arg1).slice();
+    wasm.__wbindgen_free(arg0, arg1 * 4, 4);
+    // generic import `logNestedImplTrait`: [Vector(U32)] -> Unit
+    logNestedImplTrait(v0);
+}
+export function __wbindgen_generic_000000000000001c(arg0, arg1, arg2) {
     let v0;
     if (arg0 !== 0) {
         v0 = Array.from(getArrayU16FromWasm0(arg0, arg1));
@@ -233,7 +261,7 @@ export function __wbindgen_generic_0000000000000016(arg0, arg1, arg2) {
     // generic import `logOptSlice`: [Option(Ref(Vector(U16))), U32] -> Unit
     logOptSlice(v0, arg2 >>> 0);
 }
-export function __wbindgen_generic_0000000000000017(arg0, arg1, arg2) {
+export function __wbindgen_generic_000000000000001d(arg0, arg1, arg2) {
     let v0;
     if (arg0 !== 0) {
         v0 = getArrayJsValueFromWasm0(arg0, arg1);
@@ -242,35 +270,43 @@ export function __wbindgen_generic_0000000000000017(arg0, arg1, arg2) {
     // generic import `logOptStrSlice`: [Option(Ref(Vector(NamedExternref("string")))), U32] -> Unit
     logOptStrSlice(v0, arg2 >>> 0);
 }
-export function __wbindgen_generic_0000000000000018(arg0) {
+export function __wbindgen_generic_000000000000001e(arg0) {
     // generic import `logRef`: [Ref(Externref)] -> Unit
     logRef(arg0);
 }
-export function __wbindgen_generic_0000000000000019(arg0, arg1, arg2) {
+export function __wbindgen_generic_000000000000001f(arg0) {
+    // generic import `logRef`: [Ref(F64)] -> Unit
+    logRef(arg0);
+}
+export function __wbindgen_generic_0000000000000020(arg0) {
+    // generic import `logRef`: [Ref(U32)] -> Unit
+    logRef(arg0 >>> 0);
+}
+export function __wbindgen_generic_0000000000000021(arg0, arg1, arg2) {
     var v0 = Array.from(getArrayU16FromWasm0(arg0, arg1));
     // generic import `logSlice`: [Ref(Vector(U16)), F64] -> Unit
     logSlice(v0, arg2);
 }
-export function __wbindgen_generic_000000000000001a(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000022(arg0, arg1, arg2) {
     var v0 = Array.from(getArrayU16FromWasm0(arg0, arg1));
     // generic import `logSlice`: [Ref(Vector(U16)), U32] -> Unit
     logSlice(v0, arg2 >>> 0);
 }
-export function __wbindgen_generic_000000000000001b(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000023(arg0, arg1, arg2) {
     var v0 = getArrayJsValueFromWasm0(arg0, arg1);
     wasm.__wbindgen_free(arg0, arg1 * 4, 4);
     // generic import `logStrSlice`: [Ref(Vector(NamedExternref("string"))), U32] -> Unit
     logStrSlice(v0, arg2 >>> 0);
 }
-export function __wbindgen_generic_000000000000001c(arg0) {
+export function __wbindgen_generic_0000000000000024(arg0) {
     // generic import `log`: [F64] -> Unit
     log(arg0);
 }
-export function __wbindgen_generic_000000000000001d(arg0, arg1) {
+export function __wbindgen_generic_0000000000000025(arg0, arg1) {
     // generic import `log`: [Ref(String)] -> Unit
     log(getStringFromWasm0(arg0, arg1));
 }
-export function __wbindgen_generic_000000000000001e(arg0, arg1) {
+export function __wbindgen_generic_0000000000000026(arg0, arg1) {
     let deferred0_0;
     let deferred0_1;
     try {
@@ -282,125 +318,141 @@ export function __wbindgen_generic_000000000000001e(arg0, arg1) {
         wasm.__wbindgen_free(deferred0_0, deferred0_1, 1);
     }
 }
-export function __wbindgen_generic_000000000000001f(arg0) {
+export function __wbindgen_generic_0000000000000027(arg0) {
     // generic import `log`: [U32] -> Unit
     log(arg0 >>> 0);
 }
-export function __wbindgen_generic_0000000000000020(arg0) {
+export function __wbindgen_generic_0000000000000028(arg0) {
     // generic import `console.log`: [U32] -> Unit
     console.log(arg0 >>> 0);
 }
-export function __wbindgen_generic_0000000000000021(arg0, arg1) {
+export function __wbindgen_generic_0000000000000029(arg0, arg1) {
+    // generic import `mixImplTrait`: [F64, U32] -> Unit
+    mixImplTrait(arg0, arg1 >>> 0);
+}
+export function __wbindgen_generic_000000000000002a(arg0, arg1) {
+    // generic import `mixImplTrait`: [U32, U32] -> Unit
+    mixImplTrait(arg0 >>> 0, arg1 >>> 0);
+}
+export function __wbindgen_generic_000000000000002b(arg0, arg1) {
     // generic import `mix`: [U32, F64] -> Unit
     mix(arg0 >>> 0, arg1);
 }
-export function __wbindgen_generic_0000000000000022(arg0, arg1) {
+export function __wbindgen_generic_000000000000002c(arg0, arg1) {
     // generic import `mix`: [U32, U32] -> Unit
     mix(arg0 >>> 0, arg1 >>> 0);
 }
-export function __wbindgen_generic_0000000000000023(arg0) {
+export function __wbindgen_generic_000000000000002d(arg0) {
     // generic import `new`: [F64] -> Externref
     const ret = new Widget(arg0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000024(arg0) {
+export function __wbindgen_generic_000000000000002e(arg0) {
     // generic import `new`: [U32] -> Externref
     const ret = new Widget(arg0 >>> 0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000025(arg0) {
+export function __wbindgen_generic_000000000000002f(arg0) {
     // generic import `of`: [F64] -> Externref
     const ret = Widget.of(arg0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000026(arg0) {
+export function __wbindgen_generic_0000000000000030(arg0) {
     // generic import `of`: [U32] -> Externref
     const ret = Widget.of(arg0 >>> 0);
     return ret;
 }
-export function __wbindgen_generic_0000000000000027(arg0, arg1) {
+export function __wbindgen_generic_0000000000000031(arg0, arg1) {
     // generic import `pair`: [U32, F64] -> Unit
     pair(arg0 >>> 0, arg1);
 }
-export function __wbindgen_generic_0000000000000028(arg0, arg1) {
+export function __wbindgen_generic_0000000000000032(arg0, arg1) {
+    // generic import `setImplTrait`: [Ref(Externref), F64] -> Unit
+    arg0.setImplTrait(arg1);
+}
+export function __wbindgen_generic_0000000000000033(arg0, arg1) {
+    // generic import `setImplTrait`: [Ref(Externref), U32] -> Unit
+    arg0.setImplTrait(arg1 >>> 0);
+}
+export function __wbindgen_generic_0000000000000034(arg0, arg1) {
     // generic import `set`: [Ref(Externref), F64] -> Unit
     arg0.set(arg1);
 }
-export function __wbindgen_generic_0000000000000029(arg0, arg1) {
+export function __wbindgen_generic_0000000000000035(arg0, arg1) {
     // generic import `set`: [Ref(Externref), U32] -> Unit
     arg0.set(arg1 >>> 0);
 }
-export function __wbindgen_generic_000000000000002a(arg0, arg1, arg2, arg3) {
+export function __wbindgen_generic_0000000000000036(arg0, arg1, arg2, arg3) {
     // generic import `set_indexed`: [Ref(Externref), Ref(String), U32] -> Unit
     arg0[getStringFromWasm0(arg1, arg2)] = arg3 >>> 0;
 }
-export function __wbindgen_generic_000000000000002b(arg0, arg1) {
+export function __wbindgen_generic_0000000000000037(arg0, arg1) {
     // generic import `set_tag`: [Ref(Externref), F64] -> Unit
     arg0.tag = arg1;
 }
-export function __wbindgen_generic_000000000000002c(arg0, arg1) {
+export function __wbindgen_generic_0000000000000038(arg0, arg1) {
     // generic import `set_value`: [Ref(Externref), U32] -> Unit
     arg0.value = arg1 >>> 0;
 }
-export function __wbindgen_generic_000000000000002d(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000039(arg0, arg1, arg2) {
     // generic import `spreadGeneric`: [U32, Ref(Slice(U32))] -> Unit
     spreadGeneric(arg0 >>> 0, ...(getArrayU32FromWasm0(arg1, arg2)));
 }
-export function __wbindgen_generic_000000000000002e(arg0, arg1) {
+export function __wbindgen_generic_000000000000003a(arg0, arg1) {
     var v0 = getArrayU32FromWasm0(arg0, arg1).slice();
     wasm.__wbindgen_free(arg0, arg1 * 4, 4);
     // generic import `sumItems`: [Vector(U32)] -> F64
     const ret = sumItems(v0);
     return ret;
 }
-export function __wbindgen_generic_000000000000002f(arg0) {
+export function __wbindgen_generic_000000000000003b(arg0) {
     // generic import `tag`: [Ref(Externref)] -> U32
     const ret = arg0.tag;
     return ret;
 }
-export function __wbindgen_generic_0000000000000030() { return handleError(function (arg0) {
+export function __wbindgen_generic_000000000000003c() { return handleError(function (arg0) {
     // generic import `tryGet`: [U32] -> F64
     const ret = tryGet(arg0 >>> 0);
     return ret;
 }, arguments); }
-export function __wbindgen_generic_0000000000000031() { return handleError(function (arg0) {
+export function __wbindgen_generic_000000000000003d() { return handleError(function (arg0) {
     // generic import `tryGet`: [U32] -> U32
     const ret = tryGet(arg0 >>> 0);
     return ret;
 }, arguments); }
-export function __wbindgen_generic_0000000000000032() { return handleError(function (arg0) {
+export function __wbindgen_generic_000000000000003e() { return handleError(function (arg0) {
     // generic import `tryLog`: [U32] -> Unit
     tryLog(arg0 >>> 0);
 }, arguments); }
-export function __wbindgen_generic_0000000000000033(arg0) {
+export function __wbindgen_generic_000000000000003f(arg0) {
     // generic import `value`: [Ref(Externref)] -> F64
     const ret = arg0.value;
     return ret;
 }
-export function __wbindgen_generic_0000000000000034(arg0) {
+export function __wbindgen_generic_0000000000000040(arg0) {
     // generic import `value`: [Ref(Externref)] -> U32
     const ret = arg0.value;
     return ret;
 }
-export function __wbindgen_generic_0000000000000035(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000041(arg0, arg1, arg2) {
     var v0 = getArrayU32FromWasm0(arg1, arg2).slice();
     wasm.__wbindgen_free(arg1, arg2 * 4, 4);
     // generic import `variadicLog`: [U32, Vector(U32)] -> Unit
     variadicLog(arg0 >>> 0, ...(v0));
 }
-export function __wbindgen_generic_0000000000000036(arg0, arg1, arg2) {
+export function __wbindgen_generic_0000000000000042(arg0, arg1, arg2) {
     try {
         var state0 = {a: arg0, b: arg1};
         var cb0 = (arg0) => {
             const a = state0.a;
             state0.a = 0;
             try {
-                return wasm_bindgen_0000000000000037___convert__closures_____invoke___u32______true_(a, state0.b, arg0);
+                return wasm_bindgen_0000000000000043___convert__closures_____invoke___u32______true_(a, state0.b, arg0);
             } finally {
                 state0.a = a;
             }
         };
-        // generic import `withCallback`: [RefMut(Function(Function { arguments: [U32], shim_idx: 62, ret: Unit, inner_ret: Some(Unit) })), U32] -> Unit
+        // generic import `withCallback`: [RefMut(Function(Function { arguments: [U32], shim_idx: 74, ret: Unit, inner_ret: Some(Unit) })), U32] -> Unit
         withCallback(cb0, arg2 >>> 0);
     } finally {
         state0.a = 0;
@@ -415,37 +467,37 @@ export function __wbindgen_init_externref_table() {
     table.set(offset + 2, true);
     table.set(offset + 3, false);
 }
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___bool__true_(arg0, arg1) {
-    const ret = wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___bool__true_(arg0, arg1);
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___bool__true_(arg0, arg1) {
+    const ret = wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___bool__true_(arg0, arg1);
     return ret !== 0;
 }
 
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___wasm_bindgen_0000000000000037___JsValue__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, arg2) {
-    const ret = wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___wasm_bindgen_0000000000000037___JsValue__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, arg2);
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___wasm_bindgen_0000000000000043___JsValue__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, arg2) {
+    const ret = wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___wasm_bindgen_0000000000000043___JsValue__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, arg2);
     if (ret[1]) {
         throw takeFromExternrefTable0(ret[0]);
     }
 }
 
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined_______true_(arg0, arg1, arg2, arg3) {
-    wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined___js_sys_0000000000000038___Function_fn_wasm_bindgen_0000000000000037___JsValue_____wasm_bindgen_0000000000000037___sys__Undefined_______true_(arg0, arg1, arg2, arg3);
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined_______true_(arg0, arg1, arg2, arg3) {
+    wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined___js_sys_0000000000000044___Function_fn_wasm_bindgen_0000000000000043___JsValue_____wasm_bindgen_0000000000000043___sys__Undefined_______true_(arg0, arg1, arg2, arg3);
 }
 
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___u32______true_(arg0, arg1, arg2) {
-    wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___u32______true_(arg0, arg1, arg2);
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___u32______true_(arg0, arg1, arg2) {
+    wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___u32______true_(arg0, arg1, arg2);
 }
 
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___u32__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, arg2) {
-    const ret = wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___u32__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, arg2);
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___u32__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, arg2) {
+    const ret = wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___u32__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, arg2);
     if (ret[1]) {
         throw takeFromExternrefTable0(ret[0]);
     }
 }
 
-function wasm_bindgen_0000000000000037___convert__closures_____invoke___alloc_000000000000003a___string__String__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, arg2) {
+function wasm_bindgen_0000000000000043___convert__closures_____invoke___alloc_0000000000000046___string__String__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, arg2) {
     const ptr0 = passStringToWasm0(arg2, wasm.__wbindgen_malloc, wasm.__wbindgen_realloc);
     const len0 = WASM_VECTOR_LEN;
-    const ret = wasm.wasm_bindgen_0000000000000037___convert__closures_____invoke___alloc_000000000000003a___string__String__core_0000000000000039___result__Result_____wasm_bindgen_0000000000000037___JsError___true_(arg0, arg1, ptr0, len0);
+    const ret = wasm.wasm_bindgen_0000000000000043___convert__closures_____invoke___alloc_0000000000000046___string__String__core_0000000000000045___result__Result_____wasm_bindgen_0000000000000043___JsError___true_(arg0, arg1, ptr0, len0);
     if (ret[1]) {
         throw takeFromExternrefTable0(ret[0]);
     }